pyproject.toml
# pyproject for baker-cli
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baker-cli"
version = "0.1.0"
description = "Super simple Docker Build Dependency Pipeline"
readme = "README.md"
authors = [
    {name = "Team IQ", email = "service@get-iq.com"}
]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/iq-company/baker-cli"

[project.scripts]
baker = "baker_cli.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"baker_cli" = [
  "templates/**",
]
